08_ChromaDB_vectorDB/requirements.txt
streamlit
chromadb
sentence-transformers
pymupdf
pillow
numpy
